pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rex"
version = "0.1.0"
description = "rex.com - Professional cryptocurrency AI trading platform"
authors = [{name = "Paul Turrell", email = "plturrell@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "flask-restx>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.28.0",
    "websocket-client>=1.5.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.10.0",
    "redis>=4.5.0",
    "celery>=5.2.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
    "schedule>=1.2.0",
    "cryptography>=41.0.0",
    "pyjwt>=2.8.0",
    "psycopg2-binary>=2.9.0",
    "gunicorn>=21.2.0",
    "web3>=6.0.0",
    "eth-account>=0.9.0",
]

[project.optional-dependencies]
ml = [
    "scikit-learn>=1.3.0",
    "tensorflow>=2.13.0",
    "torch>=2.0.0",
    "xgboost>=1.7.0",
    "lightgbm>=3.3.0",
    "optuna>=3.2.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
    "pre-commit>=3.3.0",
]
monitoring = [
    "prometheus-client>=0.17.0",
    "grafana-api>=1.0.3",
    "opentelemetry-api>=1.18.0",
    "opentelemetry-sdk>=1.18.0",
]

[project.scripts]
cryptotrading = "cryptotrading.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=src/cryptotrading",
    "--cov-report=html",
    "--cov-report=term-missing",
]